one test for all hay tests
it's useful to use the output of a simulator as input for an evaluator
during testing, without haveing to re-simulate. This requires tests to
be in one file (at least, that's the easiest config)
I've decided to omit to test hay_objective_fun, as it always segfaults.
You can still use h.evaluator.evaluate, just not
h.evaluator.evaluate_population. The latter tries to access a deleted
dendrite, resulting in a segfault. This is likely because the default
cell loaded in the MOEA_EH_gui is a GACell_v3, which is (i think?) a
stump cell?